+++ public/index.html
@@ -15,6 +15,5 @@
 	<body>
 		<noscript>You need to enable JavaScript to run this app.</noscript>
 		<div class="app" id="root"></div>
-		<div id="react-modals"></div>
 	</body>
 </html>
