+++ src/components/AuthForm/AuthTemplate.module.scss
@@ -3,8 +3,12 @@
 	padding: 80px 32px;
 	box-sizing: border-box;
 	display: flex;
+	z-index: 10;
 	flex-direction: column;
-	justify-content: left;
+	background: rgba(255, 255, 255, 1);
+	border-radius: 32px;
+	margin: 0 auto;
+	position: relative;
 	@media (max-width: 1298px) {
 		padding: 40px 20px;
 	}
@@ -12,6 +16,10 @@
 		padding: 8px 12px;
 		width: 100%;
 		height: 100%;
+		border-radius: 0px;
+	}
+	@media (max-height: 900px) {
+		margin: 15px auto;
 	}
 }
