+++ src/components/Modal/Modal.module.scss
@@ -1,38 +1,30 @@
 .modal {
 	display: flex;
-	flex-direction: column;
-	width: fit-content;
-	background: rgba(255, 255, 255, 1);
-	border-radius: 32px;
-	position: absolute;
+	align-items: center;
+	justify-content: center;
+	height: 100vh;
+	position: relative;
+	&::before {
+		content: '';
+		position: fixed;
+		top: 0;
+		left: 0;
+		right: 0;
+		bottom: 0;
+		background-color: rgba(45, 45, 45, 0.3);
+		width: 100%;
+		height: 100%;
+	}
+
+	@media (max-height: 900px) {
+		height: auto;
+	}
 	@media (max-width: 600px) {
 		width: 100%;
 		height: 100%;
 		border-radius: 0;
 	}
-}
-
-.modalWithOverlay {
-	top: 50%;
-	left: 50%;
-	transform: translate(-50%, -50%);
-}
-
-.button {
-	background-color: inherit;
-	border: 0;
-	padding: 0;
-	margin: 0;
-	cursor: pointer;
-	width: 28px;
-	height: 28px;
-	background-image: url('../../images/close_icon.svg');
-	align-self: flex-end;
-	position: absolute;
-	top: 35px;
-	right: 32px;
-
-	@media (max-width: 600px) {
-		top: 17px;
+	@media (max-width: 600px) and (max-height: 870px) {
+		height: auto;
 	}
 }
