--- src/components/ModalOverlay/ModalOverlay.module.scss
@@ -1,5 +1,0 @@
-.overlay {
-	width: 100vw;
-	height: 100vh;
-	background-color: rgba(45, 45, 45, 0.3);
-}
