--- src/components/NavPanel/NavPanel.module.scss
@@ -1,36 +1,0 @@
-.navPanel {
-	height: 65px;
-	background: #ececec;
-	box-shadow: 4px 0px 26px 0px rgba(180, 187, 215, 0.15);
-	position: fixed;
-	bottom: 0;
-	left: 0;
-	right: 0;
-	display: none;
-	@media (max-width: 992px) {
-		display: flex;
-		justify-content: center;
-		align-items: center;
-		gap: 35px;
-	}
-}
-
-.btn {
-	font-family: Inter, sans-serif;
-	border: none;
-	padding: 0;
-	font-size: 14px;
-}
-
-.content {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-
-.imgActive * {
-	fill: #550fda;
-}
-.active {
-	color: #550fda;
-}
