+++ src/pages/AuthPages/LoginPage.jsx
@@ -53,7 +53,7 @@ export default function LoginPage() {
 							localStorage.setItem('token', respons.data.auth_token); // записываем токен в localStorage
 							setUser({ ...res.data });
 							navigate('/templates');
-							console.log('auth')
+							console.log('auth');
 						});
 				} else {
 					// сообщаем пользователю об ошибке
@@ -87,7 +87,7 @@ export default function LoginPage() {
 
 	return (
 		visible && (
-			<Modal hasOverlay handleClose={handleClose}>
+			<Modal handleClose={handleClose}>
 				<AuthForm title="Вход">
 					<form className={styles.form} onSubmit={handleSubmit(onSubmit)}>
 						<InputForm
@@ -150,6 +150,11 @@ export default function LoginPage() {
 							Регистрация
 						</Link>
 					</form>
+					<button
+						onClick={handleClose}
+						className={styles.button}
+						aria-label="Close"
+					/>
 				</AuthForm>
 			</Modal>
 		)
