+++ src/pages/AuthPages/index.module.scss
@@ -186,3 +186,21 @@ $color-link-l: rgba(0, 25, 255, 1);
 	background-position: center;
 	animation: preloader-spin 1.5s linear infinite;
 }
+.button {
+	background-color: inherit;
+	border: 0;
+	padding: 0;
+	margin: 0;
+	cursor: pointer;
+	width: 28px;
+	height: 28px;
+	background-image: url('../../images/close_icon.svg');
+	align-self: flex-end;
+	position: absolute;
+	top: 35px;
+	right: 32px;
+	background-color: transparent;
+	@media (max-width: 600px) {
+		display: none;
+	}
+}
